requirements.txt
graphviz==0.8.4
grpcio==1.26.0
h5py==2.9.0
htmlmin==0.1.12
idna==2.8
imageio==2.6.1
importlib-metadata==1.3.0
imutils==0.5.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
joblib==0.14.1
json5==0.8.5
jsonargparse==2.20.0
jsonschema==3.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lightgbm==2.0.12
lime==0.1.1.35
llvmlite==0.31.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.0.2
numpy==1.16.4
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
opencv-python-headless==4.1.2.30
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
ppft==1.6.6.1
prettytable==0.7.2
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.2
pytest-pylint==0.14.1
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
resampy==0.2.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.13.0
terminado==0.8.3
tensorflow==1.14.0
testpath==0.4.4
tqdm==4.41.1
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
